back/requirements.txt
annotated-types==0.7.0
asgiref==3.10.0
Django==5.2.7
djangorestframework==3.16.1
dotenv==0.9.9
pillow==12.0.0
pydantic==2.12.3
pydantic-settings==2.11.0
pydantic_core==2.41.4
python-dotenv==1.2.1
sqlparse==0.5.3
typing-inspection==0.4.2
typing_extensions==4.15.0

# Added dependencies used by the project
django-parler>=3.1.0
parler-rest>=0.1.0
googletrans==4.0.0-rc1
django-cors-headers>=4.0.0
drf-yasg>=1.22.0
jazzmin>=2.0.0
django-extensions>=3.2.0
